requirements.txt
openai
anthropic
python-dotenv
fastapi
pydantic
uvicorn